# pkg/api/rest/openapi_types.gen.go
// Package rest provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package rest

// CategoriesDTO defines model for CategoriesDTO.
type CategoriesDTO []CategoryDTO

// CategoryDTO defines model for CategoryDTO.
type CategoryDTO struct {
	Description *string            `json:"description,omitempty"`
	Genres      *[]GenreOfCategory `json:"genres,omitempty"`
	Id          *string            `json:"id,omitempty"`
	Name        string             `json:"name"`
}

// CategoryOfGenre defines model for CategoryOfGenre.
type CategoryOfGenre struct {
	Description *string `json:"description,omitempty"`
	Id          *string `json:"id,omitempty"`
	Name        string  `json:"name"`
}

// Error defines model for Error.
type Error struct {
	Code    *string `json:"code,omitempty"`
	Message string  `json:"message"`
}

// GenreDTO defines model for GenreDTO.
type GenreDTO struct {
	Categories *[]CategoryOfGenre `json:"categories,omitempty"`
	Id         *string            `json:"id,omitempty"`
	Name       string             `json:"name"`
}

// GenreOfCategory defines model for GenreOfCategory.
type GenreOfCategory struct {
	Id   *string `json:"id,omitempty"`
	Name string  `json:"name"`
}

// GenresDTO defines model for GenresDTO.
type GenresDTO []GenreDTO

// CategoryCreateJSONBody defines parameters for CategoryCreate.
type CategoryCreateJSONBody CategoryDTO

// GenreCreateJSONBody defines parameters for GenreCreate.
type GenreCreateJSONBody CategoryDTO

// CategoryCreateRequestBody defines body for CategoryCreate for application/json ContentType.
type CategoryCreateJSONRequestBody CategoryCreateJSONBody

// GenreCreateRequestBody defines body for GenreCreate for application/json ContentType.
type GenreCreateJSONRequestBody GenreCreateJSONBody